+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 - Colored block recipes now output 7 instead of 1, to match the amount in the input (#9)
+- Fixed the Ender Quarry placing a Dirt block where it starts (#10)
 
 ### Changed